クラス形成
仮の実装から整理
